implementations/telegram/requirements.txt
python-telegram-bot>=21.0
-e ../../sdk/python
python-dotenv>=0.21.0
